Breaks up gps_pwv tests
